Extract server admin routes to own file
